+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## unreleased
+
+* Refactor datetime conversion in Parser. https://gitlab.com/egze/data_for_seo/-/merge_requests/19
+
+
 ## 0.3.0 (2020-03-30)
 
 Breaking change.